Modify iss and initial tp, could be wrong

+++ click-2.0.1/elements/xia/xstream.hh
@@ -88,8 +88,8 @@ using namespace xia;
 struct mini_tcpip
 {
 	uint16_t ti_len;
-	uint16_t ti_seq;
-	uint16_t ti_ack;
+	uint32_t ti_seq;
+	uint32_t ti_ack;
 	uint16_t ti_off;
 	uint16_t ti_flags;
 	uint16_t ti_win;
